I will run the unit tests when making trivial changes. (#15)
Followup for commit ad64ffd. The code in LdapServer.fetchOne was
checking whether the single element in a Set<LdapPerson> was an
LdapPerson, and trying to throw an IllegalArgumentException if it
wasn't. <Object> instanceof LdapPerson fails on null, but the null was
deferenced trying to construct the IllegalArgumentException, so a
NullPointerException was thrown instead, as "verified" by a unit test.
I have preserved the actual NPE instead of the intended IAE.

Clean up two tests to use ExpectedException more often and better.